fix warning introduced by a previous commit
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@26115 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/shape/shape2ogr.cpp
@@ -1306,7 +1306,7 @@ OGRErr SHPWriteOGRFeature( SHPHandle hSHP, DBFHandle hDBF,
               char szFormat[20];
               char szValue[32];
               int nFieldWidth = poFieldDefn->GetWidth();
-              sprintf(szFormat, "%%%dd", MIN(nFieldWidth, sizeof(szValue)-1));
+              sprintf(szFormat, "%%%dd", MIN(nFieldWidth, (int)sizeof(szValue)-1));
               sprintf(szValue, szFormat, poFeature->GetFieldAsInteger(iField) );
               int nStrLen = strlen(szValue);
               if( nStrLen > nFieldWidth )
